requirements.txt
Django~=4.2.7
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-extensions